requirements.txt
streamlit
streamlit-aggrid
pymysql
sqlalchemy
pandas
mysql-connector-python
plotly
bcrypt